Avoid quadratic scaling
If a message has many arguments and is built up in many chunks, the time
would scale proportionally to the product of the two, because the
entire set of arguments would be transferred back and forth between the
transient and permanent states on each call to append.

Instead, transfer only a trailing argument from permanent state to
transient state at the start.