Use std::hint::black_box in benchmarks
In theory this should stop the compiler making some optimisations that
would invalidate the benchmark (like throwing the whole thing away
because there are no side effects), but it seems to actually be doubling
the performance for parsing. I can only assume that it's causing
optimisation effort to be spent somewhere else.